Remove unused apps and added DRF filters.
In the afb settings of our API app, we made a revision to our list of installed apps. The unused applications `crispy_forms`, `tailwind`, and `crispy_tailwind` were removed from the list in order to declutter and streamline our app. Meanwhile, we added `rest_framework` and `django_filters`, the latter of which is a dependency of Django Rest Framework. These new inclusions will enhance the functionality and efficiency of our Django project, with `rest_framework` providing a powerful and flexible toolkit for building Web APIs, and `django_filters` promoting filterable data for Django models.

Signed-off-by: delano <delano@cpan.org>