Refactor routes in Vue router
In the Vue router, this revision introduces several enhancements.
Firstly, the `RouteRecordRaw` import from 'vue-router' is added. The
`privateApplicationRoute`, `publicApplicationRoute`, and
`publicLoginRoute` are introduced for meta data of respective routes.
Additionally, `publicRoutes` and `applicationRoutes` arrays were created
to separate the routes and provide clearer distinction between public
and private ones. Lastly, the construction of the `routes` variable was
modified to be an array of the concatenated `publicRoutes` and
`applicationRoutes`. Several comments and unnecessary lines were removed
in this process to enhance tidiness and clarity.

Signed-off-by: delano <delano@cpan.org>