Introduce immediate form submission on Next button
In the RegisterView.vue, we've added an immediate submission of form data when the user clicks the 'Next' button on the registration form. This introduces a significant UX improvement:

- Explicit axios import and addition of the `handleSubmit` function for asynchronously handling form data.
- Adjustment of form rules and some HTML attributes for compatibility with the new submission flow.
- Substitution of 'Next' function to the `handleSubmit` response to connect the enhanced flow.
- Inclusion of pre-emptive error handling, alerting users to issues like an occupied email address before they've completed the whole form.

This reduces the risk of a user filling out the entire form before discovering an error and prevents unnecessary repetition which can improve the user experience.

Signed-off-by: delano <delano@cpan.org>