[#66] Fix for nested NuxtLink issue
Getting to the bottom of this was taking too long (1h+) (I got to the middle somewhere). There are a number of StackOverflow threads and Github Issues related this the same error message so it seems to me like a bug in the Nuxt internals that pops up in various (non-obvious) circumstances. In our case my hunch is that the `<UHeader>` component that wraps the link in question here creates its own NuxtLink(s).

Console error:
```
[nuxt] [NuxtLink] You can't nest one <a> inside another <a>. This will cause a hydration error on client-side. You can pass the `custom` prop to take full control of the markup.
```