[#80] Refactored request form routes and components
* Moved request form component to reuse at `/requests/new`
* Updated nav links to open requests subtree by default
* Added request history and new request pages to requests nav

The request form component was extracted out of the old `/requestform` route so it can be reused at the new `/requests/new` location instead. The dashboard nav tree for requests was also updated to default to open when on any `/requests` path and nested routes were added for the request history and new request pages. Lastly the header navlink to the form was redirected from `/requestform` to `/requests/new` as well.