[NOJIRA] Fix for magic link URL path
Needs to include the trailing slash. Caddy will redirect to the correct
URL if it's missing, which may not include the query params correctly.

The redirect behaviour is specified in RFC 3986. It specifies that if a
URI contains an authority component, then the path component must either
be empty or begin with a slash ("/") character. This is why Caddy 2, in
compliance with RFC 3986, may add a trailing slash to reverse proxy
requests and perform a 308 redirect if the trailing slash is missing