[#39] Fix npm script commands for apps/ui subdir
NOTE: We now run the nuxt app commands from the root of the project. It's a natural (and intentional) consequence of moving the package.json up out of the apps/ui directory. This is a good thing, as it's a step towards making the project more consistent and easier to work with. We'll do the same for the Django app in the future.