[#134] Add contact page
Changes help text and links to be more clear and user-focused. Adds a new contact page form to allow users to easily reach out for support.

- Renamed "Help & FAQ" text to "Help & Support" for clarity
- Updated description on address field to remove unnecessary text and encourage contacting support
- Changed "Help & FAQ" link in footer to link to new "/contact" page
- Created new "/contact" page with basic contact form using email, name, message fields

This commit aims to improve the user experience around getting support by updating outdated help text and references, and adding a dedicated contact page form to make it easy for users to reach staff.