[#134] Bump next and next-auth back up to latest
Been seeing some weird cookie behaviour. Not sure if it's related but the downgrade (to align with @sidebase/next-auth) timing is about right. Every N minutes, the response includes 4 identical Set-Cookie headers for auth.token with a blank value. The server header says it's caddy but that could be a macguffin.

Signed-off-by: delano <delano@cpan.org>